chatbot/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
pypdf
python-multipart 
streamlit 